requirements.txt
kafka-python
apache-flink==1.20
pyflink
numpy
python-dotenv
pandas
six
xgboost
joblib
scikit-learn